Fixed the scroller for the photoblog.
The problem was that I had applied a callback to the last thumb in the thumbsContainer dl-list, because I thought it was this thumb that would be loaded last. Sometimes this doesn't always hold true, so I added a callback to every thumb that reloads the width of the handle, instead of doing it once.

I think the performancelosses are not so noticable. Sille me.

Signed-off-by: Erik Andersson <erik.rothoff@gmail.com>