Fixed so that the left modules are not hidden whilst in the photoblog. This commit changes some of the core ui code. Read below.
The change in ui_bottom is in the way module order is treated. It should workalmost exactly like before, but bugs do happen. And if modules are added to the ui_options-array and the ui_modules_hide is not set or not set to false they will be added whitout hiding the other modules. They will be added under the search box, and can be sorted.

Signed-off-by: Erik Andersson <erik.rothoff@gmail.com>